chore: clippy

+++ src/cuda_sniff.rs
@@ -57,7 +57,7 @@ pub fn sniff_ioctl(fd: c_int, request: c_ulong, argp: *mut c_void) -> Result<Opt
         NV_ESC_REGISTER_FD => {
             let params = argp as *const nv_ioctl_register_fd_t;
             format!("NV_ESC_REGISTER_FD fd: {}", unsafe {
-                (*params).ctl_fd as i32
+                (*params).ctl_fd
             })
         }
         NV_ESC_ALLOC_OS_EVENT => format!("NV_ESC_ALLOC_OS_EVENT"),
